+++ src/main/java/com/lifeform/main/network/pool/StatUpdate.java
@@ -2,13 +2,13 @@ package com.lifeform.main.network.pool;
 
 import com.lifeform.main.IKi;
 import com.lifeform.main.network.IConnectionManager;
-import com.lifeform.main.network.pool.PoolPacket;
+
 
 import java.io.Serializable;
 
 public class StatUpdate implements Serializable, PoolPacket {
     public long shares;
-    public long currentPPS;
+    public double currentPPS;
 
     @Override
     public void process(IKi ki, IConnectionManager connMan) {
